apprentice teacher and senior teacher now inherit teacher class
