+++ apprentice_teacher.rb
@@ -1,49 +1,26 @@
 require_relative 'high_five_and_phase_changeable'
+require_relative 'teacher'
 
-class ApprenticeTeacher
-  include HighFiveAndPhaseChangeable
+class ApprenticeTeacher < Teacher
 
-  attr_reader :salary, :target_raise
+  attr_reader :target_raise
 
   def initialize(options={})
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-    @target_raise = 800
-    @phase = 3
-  end
-
-  def set_phase(num)
     super
-    "Cool, I've always wanted to teach phase #{num}!"
+    @target_raise = 800
   end
 
-  def teach_stuff
+
+  def teach_stuff()
     response = ""
     response += "Listen, class, this is how everything works. "
     response += "*drops crazy knowledge bomb* "
     response += "... You're welcome."
     response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
+  end 
 
   def set_performance_rating(rating)
-    response = ""
-    if rating > 80
-      response = "Yay, I'm a great employee!"
-      receive_raise(@target_raise)
-    else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
-    end
-    response
+    super(rating, 80, target_raise)
   end
 
   def attend_training_session
