+++ senior_teacher.rb
@@ -1,20 +1,13 @@
 require_relative 'high_five_and_phase_changeable'
+require_relative 'teacher'
 
-class SeniorTeacher
-  include HighFiveAndPhaseChangeable
+class SeniorTeacher < Teacher
 
-  attr_reader :salary, :performance_rating, :target_raise
+  attr_reader :performance_rating
 
   def initialize(options={})
-    @phase = 3
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-    @target_raise = 1000
-  end
-
-  def set_phase(num)
     super
-    "Cool, I've always wanted to teach phase #{num}!"
+    @target_raise = 1000
   end
 
   def teach_stuff
@@ -25,25 +18,8 @@ class SeniorTeacher
     response
   end
 
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
-
   def set_performance_rating(rating)
-    response = ""
-    if rating > 90
-      response = "Yay, I'm a great employee!"
-      receive_raise(@target_raise)
-    else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
-    end
-    response
+    super(rating,90,@target_raise)
   end
 
   def lead_training_session
